+++ CHANGES.rst
@@ -11,6 +11,12 @@ New Features:
   it's really required.
   [datakurre]
 
+Fixes:
+
+- Fix issue where layout related fields could have been acquired
+  (only sectionSiteLayout can be allowed to be acquired)
+  [datakurre]
+
 
 4.0.5 (2017-02-08)
 ------------------